requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.0.4
idna==3.3
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2020.4
requests==2.27.1
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.9
whitenoise==5.2.0
django-import-export
boto3